Improve: integrate db util and backupper
