+++ src/Component/Backupper.ts
@@ -18,8 +18,12 @@
 
 import type { DataType } from "../botBase";
 
+import { http, https } from "follow-redirects";
+
 import { LogEmitter } from "../Structure";
 import Util from "../Util";
+
+const MIME_JSON = "application/json";
 
 export class BackUpper extends LogEmitter {
   private _queueModifiedGuilds:string[] = [];
@@ -28,8 +32,18 @@ export class BackUpper extends LogEmitter {
     return this.getData();
   }
 
+  /**
+   * 変更済みとしてマークされたキューを持つサーバーの一覧を返します
+   */
   get queueModifiedGuilds():Readonly<string[]>{
     return this._queueModifiedGuilds;
+  }
+
+  /**
+   * バックアップが実行可能な設定がされているかを示します。
+   */
+  get backuppable(){
+    return !!(process.env.GAS_TOKEN && process.env.GAS_URL);
   }
 
   constructor(private readonly getData:(()=>DataType)){
@@ -37,10 +51,17 @@ export class BackUpper extends LogEmitter {
     this.setTag("Backup");
   }
 
+  /**
+   * 今変更済みキューがあるとしてマークされているサーバーをリセットして、すべてマークを解除します。
+   */
   resetModifiedGuilds(){
     this._queueModifiedGuilds = [];
   }
 
+  /**
+   * 指定したサーバーIDのキューを、変更済みとしてマークします  
+   * マークされたサーバーのキューは、次回のティックにバックアップが試行されます
+   */
   addModifiedGuilds(guildId:string){
     if(!this._queueModifiedGuilds.includes(guildId)) this._queueModifiedGuilds.push(guildId);
   }
@@ -49,7 +70,7 @@ export class BackUpper extends LogEmitter {
    * 接続ステータスやキューを含む全データをサーバーにバックアップします
    */
   backupData():Promise<any>|void{
-    if(Util.db.DatabaseAPI.CanOperate){
+    if(this.backuppable){
       return this.backupQueue().then(() => this.backupStatus());
     }
   }
@@ -65,7 +86,7 @@ export class BackUpper extends LogEmitter {
       }));
       if(queue.length > 0){
         Util.logger.log("[Backup] Backing up modified queue...");
-        await Util.db.DatabaseAPI.SetQueueData(queue);
+        await this.backupQueueData(queue);
         this._queueModifiedGuilds = [];
       }else{
         Util.logger.log("[Backup] No modified queue found, skipping");
@@ -96,7 +117,7 @@ export class BackUpper extends LogEmitter {
       });
       if(speaking.length > 0){
         Util.logger.log("[Backup] Backing up modified status..");
-        await Util.db.DatabaseAPI.SetIsSpeaking(speaking);
+        await this.backupStatusData(speaking);
         this._previousStatuses = currentStatuses;
       }else{
         Util.logger.log("[Backup] No modified status found, skipping");
@@ -106,4 +127,183 @@ export class BackUpper extends LogEmitter {
       this.Log(e, "warn");
     }
   }
+
+  /**
+   * ステータスデータをサーバーから取得する
+   */
+  async getStatusFromServer(guildids:string[]){
+    if(this.backuppable){
+      const t = Util.time.timer.start("GetIsSpeking");
+      try{
+        const result = await this.requestHttp("GET", process.env.GAS_URL, {
+          token: process.env.GAS_TOKEN,
+          guildid: guildids.join(","),
+          type: "j"
+        } as requestBody, MIME_JSON);
+        if(result.status === 200){
+          return result.data as {[guildid:string]:string};
+        }else{
+          return null;
+        }
+      }
+      catch{
+        return null;
+      }
+      finally{
+        t.end();
+      }
+    }else{
+      return null;
+    }
+  }
+
+  /**
+   * キューのデータをサーバーから取得する
+   */
+  async getQueueDataFromServer(guildids:string[]){
+    if(this.backuppable){
+      const t = Util.time.timer.start("GetQueueData");
+      try{
+        const result = await this.requestHttp("GET", process.env.GAS_URL, {
+          token: process.env.GAS_TOKEN,
+          guildid: guildids.join(","),
+          type: "queue"
+        } as requestBody, MIME_JSON);
+        if(result.status === 200){
+          return result.data as {[guildid:string]:string};
+        }else return null;
+      }
+      catch(e){
+        return null;
+      }
+      finally{
+        t.end();
+      }
+    }else{
+      return null;
+    }
+  }
+
+  /**
+   * ステータス情報をサーバーへバックアップする
+   */
+  private async backupStatusData(data:{guildid:string, value:string}[]){
+    if(this.backuppable){
+      const t = Util.time.timer.start("backupStatusData");
+      const ids = data.map(d => d.guildid).join(",");
+      const rawData = {} as {[key:string]:string};
+      data.forEach(d => rawData[d.guildid] = d.value);
+      try{
+        const result = await this.requestHttp("POST", process.env.GAS_URL, {
+          token: process.env.GAS_TOKEN,
+          guildid: ids,
+          data: JSON.stringify(rawData),
+          type: "j"
+        } as requestBody, MIME_JSON);
+        if(result.status === 200){
+          return true;
+        }else{
+          return false;
+        }
+      }
+      catch{
+        return false;
+      }
+      finally{
+        t.end();
+      }
+    }else{
+      return false;
+    }
+  }
+
+  /**
+   * キューのデータをサーバーへバックアップする
+   */
+  private async backupQueueData(data:{guildid:string, queue:string}[]){
+    if(this.backuppable){
+      const t = Util.time.timer.start("SetQueueData");
+      const ids = data.map(d => d.guildid).join(",");
+      const rawData = {} as {[guildid:string]:string};
+      data.forEach(d => rawData[d.guildid] = encodeURIComponent(d.queue));
+      try{
+        const result = await this.requestHttp("POST", process.env.GAS_URL, {
+          token: process.env.GAS_TOKEN,
+          guildid: ids,
+          data: JSON.stringify(rawData),
+          type: "queue"
+        } as requestBody, MIME_JSON);
+        return result.status === 200;
+      }
+      catch{
+        return false;
+      }
+      finally{
+        t.end();
+      }
+    }else{
+      return false;
+    }
+  }
+
+  /**
+   * HTTPでデータをバックアップするユーティリティメソッド
+   */
+  private async requestHttp(method:"GET"|"POST", url:string, data?:requestBody, mimeType?:string){
+    return new Promise<postResult>((resolve, reject) => {
+      if(method === "GET"){
+        url += "?" + (Object.keys(data) as (keyof requestBody)[]).map(k => encodeURIComponent(k) + "=" + encodeURIComponent(data[k])).join("&");
+      }
+      const durl = new URL(url);
+      const opt = {
+        protocol: durl.protocol,
+        hostname: durl.hostname,
+        port: durl.port,
+        path: durl.pathname + durl.search,
+        method: method,
+      } as {[key:string]:any};
+      if(mimeType){
+        opt.headers = {
+          "Content-Type": mimeType
+        };
+      }
+      const httpLibs = {
+        "http:": http,
+        "https:": https,
+      } as {[proto:string]:(typeof http|typeof https)};
+      const req = httpLibs[durl.protocol].request(opt, (res) => {
+        const bufs = [] as Buffer[];
+        res.on("data", chunk => bufs.push(chunk));
+        res.on("end", ()=> {
+          try{
+            const parsed = JSON.parse(Buffer.concat(bufs).toString("utf-8")) as postResult;
+            if(parsed.data) Object.keys(parsed.data).forEach(k => parsed.data[k] = decodeURIComponent(parsed.data[k]));
+            resolve(parsed);
+          }
+          catch(e){
+            reject(e);
+          }
+        });
+        res.on("error", ()=>reject());
+      });
+      if(method === "POST"){
+        req.end(JSON.stringify(data));
+      }else{
+        req.end();
+      }
+    });
+  }
 }
+
+type getResult = {
+  status: 200|404,
+};
+type postResult = getResult & {
+  data:any,
+};
+type requestBody = {
+  token:string,
+  guildid:string,
+  data?:any,
+  type:"queue"|"j",
+};
